requirements.txt
Flask==3.1.2
Jinja2==3.1.6
Werkzeug==3.1.3
gunicorn==21.2.0
python-dotenv==1.0.0
